+++ app/src/main/AndroidManifest.xml
@@ -1,17 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="cecs492.com.decisionator" >
+    package="com.csulb.decisionator.decisionator">
 
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme" >
-        <activity
-            android:name=".Decisionator_Login"
-            android:label="@string/app_name"
-            android:theme="@style/AppTheme.NoActionBar" >
+        android:theme="@style/AppTheme">
+        <activity android:name=".FacebookLogin">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
