+++ app/src/main/java/com/csulb/decisionator/decisionator/FacebookLogin.java
@@ -1,0 +1,13 @@
+package com.csulb.decisionator.decisionator;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class FacebookLogin extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_facebook_login);
+    }
+}
