+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,3 @@
 <resources>
     <string name="app_name">Decisionator</string>
-
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
-
-    <string name="action_settings">Settings</string>
 </resources>
